Extract TestCluster interface
