+++ bundle-workflow/python/test.py
@@ -4,7 +4,7 @@ import os
 import argparse
 from manifests.bundle_manifest import BundleManifest
 from git.git_repository import GitRepository
-from test_workflow.test_cluster import LocalTestCluster
+from test_workflow.local_test_cluster import LocalTestCluster
 from test_workflow.integ_test_suite import IntegTestSuite
 from paths.script_finder import ScriptFinder
 from system.temporary_directory import TemporaryDirectory
